Make list headers match page headers
